backend/api/requirements.txt
fastapi       # Фреймворк для API
uvicorn       # ASGI-сервер для запуска FastAPI
minio         # Клиентская библиотека MinIO (S3 совместимое хранилище)
kafka-python  # Библиотека для работы с Apache Kafka
elasticsearch      
requests 
prometheus_fastapi_instrumentator
python-jose[cryptography]
pyjwt